refactor(api): port tatoeba, tts, extract-url to Hono and proxy them
Three external/compute routes with no DB needs:
- tatoeba: node:https search (IPv4-forced) against tatoeba.org, af→eng.
- tts: Google Cloud TTS (base64 audio), 503 fallback when no API key.
- extract-url: @mozilla/readability + linkedom → markdown article;
  adds those two deps to the api package (they run fine under Bun).

Registered in index.ts (auth SCOPE_MAP already covered all three). The
Next routes become proxies. Verified through the proxy: a live tatoeba
search, extract-url on a real page, and the tts no-key/no-text paths.
