+++ api/package.json
@@ -10,9 +10,11 @@
   "dependencies": {
     "@anthropic-ai/claude-agent-sdk": "^0.2.114",
     "@anthropic-ai/sdk": "^0.78.0",
+    "@mozilla/readability": "^0.6.0",
     "@sentry/node": "^10",
     "adm-zip": "^0.5.17",
     "hono": "^4.12.25",
+    "linkedom": "^0.18.12",
     "uuid": "^14.0.0",
     "yaml": "^2.8.3"
   },
