+++ api/src/index.ts
@@ -7,6 +7,9 @@ import collections from './routes/collections';
 import groups from './routes/groups';
 import knownWords from './routes/known-words';
 import studyPing from './routes/study-ping';
+import tatoeba from './routes/tatoeba';
+import tts from './routes/tts';
+import extractUrl from './routes/extract-url';
 import lessons from './routes/lessons';
 import vocab from './routes/vocab';
 import cloze from './routes/cloze';
@@ -32,6 +35,9 @@ app.route('/api/collections', collections);
 app.route('/api/groups', groups);
 app.route('/api/known-words', knownWords);
 app.route('/api/study-ping', studyPing);
+app.route('/api/tatoeba', tatoeba);
+app.route('/api/tts', tts);
+app.route('/api/extract-url', extractUrl);
 app.route('/api/lessons', lessons);
 app.route('/api/vocab', vocab);
 app.route('/api/cloze', cloze);
